% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNrich.R
\name{BN_struct}
\alias{BN_struct}
\title{Construct Bayesian networks structures}
\usage{
BN_struct(mapkG1)
}
\arguments{
\item{mapkG1}{A list contains unified signaling pathways}
}
\value{
A list contains Bayesian networks structures
}
\description{
Construct BNs structures using unified signaling pathways
}
