% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNrich.R
\name{BNrich}
\alias{BNrich}
\title{Analysis of significant final BNs}
\usage{
BNrich(Ttest_results, fdr.value = 0.05, pathway.id1, PathName_final)
}
\arguments{
\item{Ttest_results}{A data frame contains T-test results for all parameters}

\item{fdr.value}{A numeric threshold to determine significant parameters}

\item{pathway.id1}{A vector contains modified KEEG pathway IDs}

\item{PathName_final}{A data frame contains is IDs and names of KEEG pathways}
}
\value{
A data frame contains fisher test results for any final pathways
}
\description{
Fisher's exact test applied to PEA on final BNs
}
