\name{dataD}
\alias{dataD}
\docType{data}
\title{
dataD
}
\description{
The example data extracted from a part of `GSE47756` dataset (patients), the gene expression data from colorectal cancer study
}

\format{
  A data frame with 3900 observations on the following 21 variables.
  \describe{
    \item{\code{D1}}{a numeric vector which shows gene expression data related to patients 1}
    \item{\code{D2}}{a numeric vector which shows gene expression data related to patients 2}
    \item{\code{D3}}{a numeric vector which shows gene expression data related to patients 3}
    \item{\code{D4}}{a numeric vector which shows gene expression data related to patients 4}
    \item{\code{D5}}{a numeric vector which shows gene expression data related to patients 5}
    \item{\code{D6}}{a numeric vector which shows gene expression data related to patients 6}
    \item{\code{D7}}{a numeric vector which shows gene expression data related to patients 7}
    \item{\code{D8}}{a numeric vector which shows gene expression data related to patients 8}
    \item{\code{D9}}{a numeric vector which shows gene expression data related to patients 9}
    \item{\code{D10}}{a numeric vector which shows gene expression data related to patients 10}
    \item{\code{D11}}{a numeric vector which shows gene expression data related to patients 11}
    \item{\code{D12}}{a numeric vector which shows gene expression data related to patients 12}
    \item{\code{D13}}{a numeric vector which shows gene expression data related to patients 13}
    \item{\code{D14}}{a numeric vector which shows gene expression data related to patients 14}
    \item{\code{D15}}{a numeric vector which shows gene expression data related to patients 15}
    \item{\code{D16}}{a numeric vector which shows gene expression data related to patients 16}
    \item{\code{D17}}{a numeric vector which shows gene expression data related to patients 17}
    \item{\code{D18}}{a numeric vector which shows gene expression data related to patients 18}
    \item{\code{D19}}{a numeric vector which shows gene expression data related to patients 19}
    \item{\code{D20}}{a numeric vector which shows gene expression data related to patients 20}
    \item{\code{D21}}{a numeric vector which shows gene expression data related to patients 21}
  }
}

\source{
The example data extracted from a part of `GSE47756` dataset (patients), the gene expression data from colorectal cancer study
}

\examples{
\dontrun{
unify_path()
}}
\keyword{datasets}
