\name{dataH}
\alias{dataH}
\docType{data}
\title{
dataH
}
\description{
The example data extracted from a part of `GSE47756` dataset (healthy control subjects), the gene expression data from colorectal cancer study
}

\alias{unify_path(dataH,dataD,mapkG = mapkG,pathway.id = pathway.id)}
\format{
  A data frame with 3900 observations on the following 21 variables.
  \describe{
    \item{\code{H1}}{a numeric vector which shows gene expression data related to control 1}
    \item{\code{H2}}{a numeric vector which shows gene expression data related to control 2}
    \item{\code{H3}}{a numeric vector which shows gene expression data related to control 3}
    \item{\code{H4}}{a numeric vector which shows gene expression data related to control 4}
    \item{\code{H5}}{a numeric vector which shows gene expression data related to control 5}
    \item{\code{H6}}{a numeric vector which shows gene expression data related to control 6}
    \item{\code{H7}}{a numeric vector which shows gene expression data related to control 7}
    \item{\code{H8}}{a numeric vector which shows gene expression data related to control 8}
    \item{\code{H9}}{a numeric vector which shows gene expression data related to control 9}
    \item{\code{H10}}{a numeric vector which shows gene expression data related to control 10}
    \item{\code{H11}}{a numeric vector which shows gene expression data related to control 11}
    \item{\code{H12}}{a numeric vector which shows gene expression data related to control 12}
    \item{\code{H13}}{a numeric vector which shows gene expression data related to control 13}
    \item{\code{H14}}{a numeric vector which shows gene expression data related to control 14}
    \item{\code{H15}}{a numeric vector which shows gene expression data related to control 15}
    \item{\code{H16}}{a numeric vector which shows gene expression data related to control 16}
    \item{\code{H17}}{a numeric vector which shows gene expression data related to control 17}
    \item{\code{H18}}{a numeric vector which shows gene expression data related to control 18}
    \item{\code{H19}}{a numeric vector which shows gene expression data related to control 19}
    \item{\code{H20}}{a numeric vector which shows gene expression data related to control 20}
    \item{\code{H21}}{a numeric vector which shows gene expression data related to control 21}
  }
}

\source{
The example data extracted from a part of `GSE47756` dataset (healthy control subjects), the gene expression data from colorectal cancer study
}

\examples{
\dontrun{
unify_path()
}}
\keyword{datasets}
