% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNrich.R
\name{esti_par}
\alias{esti_par}
\title{Estimate parameters of BNs in control and disease states}
\usage{
esti_par(BN_H, BN_D, data_h, data_d)
}
\arguments{
\item{BN_H}{A list contains simplified BNs structures for control objects}

\item{BN_D}{A list contains simplified BNs structures for disease objects}

\item{data_h}{A list contains data frames related to control objects for any BN}

\item{data_d}{A list contains data frames related to disease objects for any BN}
}
\value{
A listcontains four lists BN_h, BN_d, coef_h and coef_d
}
\description{
Estimate parameters of BNs in control and disease states
}
