% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNrich.R
\name{fetch_data_file}
\alias{fetch_data_file}
\title{Download data file}
\usage{
fetch_data_file(destfile, verbose = FALSE)
}
\arguments{
\item{destfile}{A directory in user's own computer for save preprocessed data file}

\item{verbose}{A logical argument to show verbose results}
}
\value{
A list contain mapkG, PathName_final and pathway.id. The mapkG is a list contains imported 187 preprocessed signaling pathways, PathName_final is a data.frame includes names and IDs of all 187 pathways and pathway.id is a character vector of pathways IDs
}
\description{
Download necessary data file to start BNrich
}
\examples{
\donttest{
destfile = tempfile("files", fileext = ".rda")
files <- fetch_data_file()
load(destfile)
}
}
