% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNrich.R
\name{LASSO_BN}
\alias{LASSO_BN}
\title{LASSO regression}
\usage{
LASSO_BN(BN, data_h, data_d)
}
\arguments{
\item{BN}{A list of Bayesian networks achieved by BN_struct function}

\item{data_h}{A list contains data frames related to control objects}

\item{data_d}{A list contains data frames related to disease objects}
}
\value{
A list contains two lists.BN_H and BN_D are simplified BNs
}
\description{
LASSO regression – second step of simplification of BNs structures
}
