% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNrich.R
\name{parm_Ttest}
\alias{parm_Ttest}
\title{Testing the equality regression coefficients}
\usage{
parm_Ttest(
  data_h,
  coef_h,
  BN_h,
  data_d,
  coef_d,
  BN_d,
  var_mat_Bh,
  var_mat_Bd,
  pathway.id1
)
}
\arguments{
\item{data_h}{A list contains data frames related to control objects for any BN}

\item{coef_h}{A list contains parameters of BN_h}

\item{BN_h}{A list contains BNs learned by control objects data}

\item{data_d}{A list contains data frames related to disease objects for any BN}

\item{coef_d}{A list contains parameters of BN_d}

\item{BN_d}{A list contains BNs learned by disease objects data}

\item{var_mat_Bh}{A list contains covariance matrixes for any node of BN_h}

\item{var_mat_Bd}{A list contains covariance matrixes for any node of BN_d}

\item{pathway.id1}{A vector contains modified KEEG pathway IDs}
}
\value{
A data frame contains T-test results for all parameters in final BNs
}
\description{
t-test for equality the corresponging parameters in any BN
}
\examples{
\dontrun{
parm_Ttest(data_h,coef_h,BN_h,data_d,coef_d,BN_d,var_mat_Bh,var_mat_Bd, pathway.id1)
}
}
