% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNrich.R
\name{start_files}
\alias{start_files}
\title{Download necessary files to start BNrich}
\usage{
start_files()
}
\value{
A list contain mapkG, PathName_final and pathway.id
mapkG is a list contains imported 187 preprocessed signaling pathways
PathName_final is a data.frame includes names and IDs of all 187 pathways
pathway.id is a character vector of pathways IDs
}
\description{
Download necessary files to start BNrich
}
\examples{
\dontrun{
start_files()
}
}
