% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNrich.R
\name{var_mat}
\alias{var_mat}
\title{Estimate variance-covariance matrixes for any parameters of BNs}
\usage{
var_mat(data_h, coef_h, BN_h, data_d, coef_d, BN_d)
}
\arguments{
\item{data_h}{A list contains data frames related to control objects for any BN}

\item{coef_h}{A lists of parameters of BN_h achieved}

\item{BN_h}{A list of BNs learned by control objects data}

\item{data_d}{A list contains data frames related to disease objects for any BN}

\item{coef_d}{A lists of parameters of BN_d}

\item{BN_d}{A list of BNs learned by disease objects data}
}
\value{
A listcontains two lists var_mat_Bh and var_mat_Bd
}
\description{
Estimate variance-covariance matrixes for any parameters of
}
\examples{
var_mat(data_h,coef_h,BN_h,data_d,coef_d,BN_d)
}
