% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNrich.R
\name{BN_struct}
\alias{BN_struct}
\title{Construct Bayesian networks structures}
\usage{
BN_struct(mapkG1)
}
\arguments{
\item{mapkG1}{A list contains unified signaling pathways}
}
\value{
A list contains Bayesian networks structures
}
\description{
Construct BNs structures using unified signaling pathways
}
\examples{
#All the 187 preprocessed signaling pathways can be entered in analysis by fetch_data_file().
#But here you enter a subset of those pathways to see how this package works.
files <- system.file("extdata", "test_files_to_start.RData", package = "BNrich", mustWork = TRUE)
load(files)
Data <- system.file("extdata", "Test_DATA.RData", package = "BNrich", mustWork = TRUE)
load(Data)
uni_Result <- unify_path(dataH, dataD, MapkG = sub_mapkG, Pathway.id = path.id)
M1 <- uni_Result$mapkG1
BN <- BN_struct(M1)
}
