% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNrich.R
\name{LASSO_BN}
\alias{LASSO_BN}
\title{LASSO regression}
\usage{
LASSO_BN(BN, data_h, data_d)
}
\arguments{
\item{BN}{A list of Bayesian networks achieved by BN_struct function}

\item{data_h}{A list contains data frames related to control objects}

\item{data_d}{A list contains data frames related to disease objects}
}
\value{
A list contains two lists.BN_h and BN_d are simplified BNs
}
\description{
LASSO regression – second step of simplification of BNs structures
}
\examples{
#All the 187 preprocessed signaling pathways can be entered in analysis by fetch_data_file().
#But here you enter a subset of those pathways to see how this package works.
files <- system.file("extdata", "test_files_to_start.RData", package = "BNrich", mustWork = TRUE)
load(files)
Data <- system.file("extdata", "Test_DATA.RData", package = "BNrich", mustWork = TRUE)
load(Data)
uni_Result <- unify_path(dataH, dataD, MapkG = sub_mapkG, Pathway.id = path.id)
M1 <- uni_Result$mapkG1
BN <- BN_struct(M1)
data_h1 <- uni_Result$data_h
data_d1 <- uni_Result$data_d
LASSO_Result <- LASSO_BN(BN = BN , data_h = data_h1 , data_d = data_d1)
}
