% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNrich.R
\name{unify_path}
\alias{unify_path}
\title{Simplification networks -- applied to unifying nodes}
\usage{
unify_path(dataH = dataH, dataD = dataD, mapkG = mapkG,
  pathway.id = pathway.id)
}
\arguments{
\item{dataH}{A data frame contains (healthy) control objects data}

\item{dataD}{A data frame contains disease objects data}

\item{mapkG}{A list contains imported 187 signaling pathways}

\item{pathway.id}{A vector contains 187 KEEG pathway IDs}
}
\value{
A list contain data_h,data_d,mapkG1 and pathway.id1
}
\description{
Unifying nodes based imported signaling pathways and GE data
}
\examples{
unify_path()
}
